// packages/codegen/src/output/database/src/mongoose/interfaces/iProcessTrackingCreateInput.ts
// THIS CODE WAS AUTOMATICALLY GENERATED
import {databaseTypes} from 'types';
import {Types as mongooseTypes} from 'mongoose';

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface IProcessTrackingCreateInput
  extends Omit<databaseTypes.IProcessTracking, '_id' | 'createdAt' | 'updatedAt'> {}